# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NewsBrief Core API
  version: 1.0.0
  description: API for authentication, email verification, and Google OAuth.
servers:
  - url: https://news-brief-core-api-excr.onrender.com
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/google/callback:
    get:
      summary: Google OAuth callback
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
        - in: query
          name: platform
          schema:
            type: string
            enum: [web, mobile]
      responses:
        '200':
          description: Tokens for mobile
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '302':
          description: Redirect with tokens in fragment
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...